# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Public Service Assistant API",
    "version": "1.0.0",
    "description": "API documentation for the Public Service Assistant project (using RAG)"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Créer une notification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification créée"
          }
        }
      },
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Lister les notifications d'un utilisateur",
        "responses": {
          "200": {
            "description": "Liste des notifications"
          }
        }
      }
    },
    "/api/notifications/{id}/read": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Marquer une notification comme lue",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification mise à jour"
          }
        }
      }
    },
    "/api/notifications/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Supprimer une notification",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notification supprimée"
          }
        }
      }
    }
  },
...
}